% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{as.data.table.jsdiff_changes}
\alias{as.data.table.jsdiff_changes}
\title{Coerce jsdiff change objects to a plain data.table / data.frame}
\usage{
\method{as.data.table}{jsdiff_changes}(x, ...)
}
\arguments{
\item{x}{A \code{jsdiff_changes} object.}

\item{...}{Unused.}
}
\value{
A \code{data.table} with columns \code{value}, \code{added}, \code{removed}, \code{count}.
}
\description{
Coerce jsdiff change objects to a plain data.table / data.frame
}
