% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_changes_to_dmp}
\alias{convert_changes_to_dmp}
\title{Convert change objects to the Google diff-match-patch format}
\usage{
convert_changes_to_dmp(changes)
}
\arguments{
\item{changes}{A \code{jsdiff_changes} object.}
}
\value{
A list of length-2 lists \code{list(operation, value)}.
}
\description{
Returns a list of \code{c(operation, value)} pairs, where operation is \code{1}
(insertion), \code{-1} (deletion), or \code{0} (equality).
}
