% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{diff_css_default}
\alias{diff_css_default}
\title{Default CSS for \code{\link[=diff_to_html]{diff_to_html()}} output}
\usage{
diff_css_default()
}
\value{
A length-1 character string of CSS rules.
}
\description{
Default CSS for \code{\link[=diff_to_html]{diff_to_html()}} output
}
