% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{is_changes}
\alias{is_changes}
\title{Are these jsdiff change objects?}
\usage{
is_changes(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Are these jsdiff change objects?
}
