% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-line-endings.R
\name{is_unix}
\alias{is_unix}
\title{Does a patch consistently use Unix line endings?}
\usage{
is_unix(patch)
}
\arguments{
\item{patch}{A \code{jsdiff_patch} or list of them.}
}
\value{
A logical scalar.
}
\description{
Does a patch consistently use Unix line endings?
}
