% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-line-endings.R
\name{is_win}
\alias{is_win}
\title{Does a patch consistently use Windows line endings?}
\usage{
is_win(patch)
}
\arguments{
\item{patch}{A \code{jsdiff_patch} or list of them.}
}
\value{
A logical scalar.
}
\description{
Does a patch consistently use Windows line endings?
}
