% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{jsdiff-aliases}
\alias{jsdiff-aliases}
\alias{diffChars}
\alias{diffWords}
\alias{diffWordsWithSpace}
\alias{diffLines}
\alias{diffTrimmedLines}
\alias{diffSentences}
\alias{diffCss}
\alias{diffJson}
\alias{diffArrays}
\alias{structuredPatch}
\alias{formatPatch}
\alias{createTwoFilesPatch}
\alias{createPatch}
\alias{parsePatch}
\alias{applyPatch}
\alias{applyPatches}
\alias{reversePatch}
\alias{convertChangesToXML}
\alias{convertChangesToDMP}
\title{jsdiff-style camelCase aliases}
\usage{
diffChars(...)

diffWords(...)

diffWordsWithSpace(...)

diffLines(...)

diffTrimmedLines(...)

diffSentences(...)

diffCss(...)

diffJson(...)

diffArrays(...)

structuredPatch(...)

formatPatch(...)

createTwoFilesPatch(...)

createPatch(...)

parsePatch(...)

applyPatch(...)

applyPatches(...)

reversePatch(...)

convertChangesToXML(...)

convertChangesToDMP(...)
}
\arguments{
\item{...}{Passed through to the underlying snake_case function.}
}
\value{
The same value as the underlying function.
}
\description{
These thin wrappers expose the original jsdiff names. See the snake_case
functions they forward to for documentation and arguments.
}
