% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-parse.R
\name{parse_patch}
\alias{parse_patch}
\title{Parse a unified diff string into structured patch objects}
\usage{
parse_patch(uni_diff)
}
\arguments{
\item{uni_diff}{A length-1 character string in unified diff format.}
}
\value{
A list of \code{jsdiff_patch} objects.
}
\description{
Parse a unified diff string into structured patch objects
}
