% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-reverse.R
\name{reverse_patch}
\alias{reverse_patch}
\title{Reverse a structured patch}
\usage{
reverse_patch(patch)
}
\arguments{
\item{patch}{A \code{jsdiff_patch} object or a list of them.}
}
\value{
A reversed \code{jsdiff_patch} (or list of them).
}
\description{
Returns a patch that undoes the supplied patch.
}
