% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-line-endings.R
\name{unix_to_win}
\alias{unix_to_win}
\title{Convert a patch's line endings from Unix to Windows}
\usage{
unix_to_win(patch)
}
\arguments{
\item{patch}{A \code{jsdiff_patch} or list of them.}
}
\value{
The converted patch.
}
\description{
Convert a patch's line endings from Unix to Windows
}
