% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-line-endings.R
\name{win_to_unix}
\alias{win_to_unix}
\title{Convert a patch's line endings from Windows to Unix}
\usage{
win_to_unix(patch)
}
\arguments{
\item{patch}{A \code{jsdiff_patch} or list of them.}
}
\value{
The converted patch.
}
\description{
Convert a patch's line endings from Windows to Unix
}
