% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdiffr-package.R
\docType{package}
\name{jsdiffr-package}
\alias{jsdiffr}
\alias{jsdiffr-package}
\title{jsdiffr: Text and Sequence Diffing Ported from 'jsdiff'}
\description{
A faithful R port of the 'jsdiff' JavaScript library. Computes differences between two pieces of text or two sequences at the level of characters, words, lines, sentences, CSS tokens, 'JSON' values, or arbitrary arrays, using Myers' O(ND) difference algorithm. The core diff loop is implemented in C++ for speed on large inputs. Also provides the full 'jsdiff' patch suite (create, apply, parse, and reverse unified diff patches), converters to XML and the Google diff-match-patch format, and helpers for rendering diffs as coloured console output or HTML for use in 'shiny' apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KyleHaynes/jsdiffr}
  \item Report bugs at \url{https://github.com/KyleHaynes/jsdiffr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Haynes \email{kyle@kylehaynes.com.au}

Other contributors:
\itemize{
  \item Kevin Decker (Author of the original 'jsdiff' JavaScript library (https://github.com/kpdecker/jsdiff)) [contributor, copyright holder]
}

}
\keyword{internal}
